fix: fulfill empty embedding for trashed docs
