+++ packages/backend/server/src/plugins/copilot/embedding/types.ts
@@ -98,8 +98,6 @@ export type Chunk = {
   content: string;
 };
 
-export const EMBEDDING_DIMENSIONS = 1024;
-
 export abstract class EmbeddingClient {
   async configured() {
     return true;
